app/requirements.txt
asgiref==3.8.1
boto3==1.35.5
botocore==1.35.5
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
pillow==10.4.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
